@@ -10,18 +10,18 @@
     "url": "r.palacios.ca"
   },
   "dependencies": {
-    "about-window": "^1.7.1",
+    "about-window": "1.8.0",
     "rc-slider": "8.3.1",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-redux": "^5.0.5",
+    "react": "16.0.0",
+    "react-dom": "16.0.0",
+    "react-redux": "5.0.6",
     "redux": "^3.7.2",
-    "uuid": "^3.0.1"
+    "uuid": "3.1.0"
   },
   "devDependencies": {
-    "electron": "^1.6.2",
-    "eslint-plugin-react": "^7.1.0",
-    "react-scripts": "0.9.5"
+    "electron": "1.7.8",
+    "eslint-plugin-react": "7.4.0",
+    "react-scripts": "1.0.14"
   },
   "homepage": "./",
   "main": "src/electron-starter.js",
