@@ -67,4 +67,4 @@ const mapDispatchToProps = (dispatch) => {
   };
 };
 
-module.exports = connect(mapStateToProps, mapDispatchToProps)(NodePanelHeader);
+export default connect(mapStateToProps, mapDispatchToProps)(NodePanelHeader);
