@@ -178,4 +178,4 @@ const mapDispatchToProps = (dispatch) => {
   };
 };
 
-module.exports = connect(mapStateToProps, mapDispatchToProps)(SynthNodePanel);
+export default connect(mapStateToProps, mapDispatchToProps)(SynthNodePanel);
