@@ -148,4 +148,4 @@ class Channel extends React.Component {
   }
 }
 
-module.exports = Channel;
+export default Channel;
