@@ -54,4 +54,4 @@ class Slider extends React.Component {
   }
 }
 
-module.exports = Slider;
+export default Slider;
