@@ -24,4 +24,4 @@ class Panel extends React.Component {
   }
 }
 
-module.exports = Panel;
+export default Panel;
