@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   octaves: [-2, -1, 0, 1, 2, 3, 4, 5, 6, 7],
   notes: [
     'C',
