@@ -65,4 +65,4 @@ class Filter {
   }
 }
 
-module.exports = Filter;
+export default Filter;
