@@ -133,4 +133,4 @@ class Node {
   }
 }
 
-module.exports = Node;
+export default Node;
